Increased max_iterations
Although 40 labels seems high enough for most use cases (e.g. reverse resolution of a IPv6 address) our project ran into issues with a domain name with 45 labels. To avoid problems in the future, I suggest to allow for the maximum number of labels allowed for a domain name. Since the function stops immediately when a shorter DNS name is processed, this should not have any impact for other users of this library.